Vision Transformer
This was a vision transformer deep learning model which was originally created for fruit segmentation research paper. It is being saved here so that it can be used in the future later. 

Remember, it performs both training and testing operations. You just have to update training and testing datasets if u are working with some other data in the future and fine tune the hyperparameters as per your requirement...Happy Coding :-)